# openapi.yaml
openapi: 3.0.3
info:
  title: Arcadia Finance API
  version: "1.0.0"
  description: |
    REST API for the Arcadia Finance demo banking application.

    WARNING: This application is intentionally vulnerable (SQL injection,
    plain-text passwords, verbose errors) for F5 WAF/security demonstration
    purposes. NEVER deploy in production.

    ## Authentication

    All private endpoints require a valid JWT Bearer token.
    Tokens are issued by the app itself (HS256, 8-hour expiry).

    Workflow:
      1. POST /api/token  with username + password -> receive access_token
      2. Every request:   Authorization: Bearer <access_token>

  contact:
    name: Arcadia Finance Demo
  license:
    name: Internal Demo - Not for production use
servers:
  - url: http://localhost:8080
    description: Local Docker Compose stack
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by POST /api/token or POST /api/login. Valid 8 hours.
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: alice
        password:
          type: string
          format: password
          example: alice123
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          description: Token lifetime in seconds
          example: 28800
        user:
          $ref: "#/components/schemas/User"
    LoginResponse:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            access_token:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            token_type:
              type: string
              example: bearer
            expires_in:
              type: integer
              example: 28800
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Not authenticated
        sql:
          type: string
          description: Intentional SQLi surface -- raw SQL returned in error responses
          example: "SELECT * FROM users WHERE username = '' OR '1'='1' --"
...
paths:
  /api/token:
    post:
      tags: [Authentication]
      summary: Obtain a JWT (API clients -- Postman, curl)
      description: |
        Exchange username + password for a signed JWT Bearer token.
        Does NOT set a session cookie (use POST /api/login for the browser).
        NOTE: Intentionally vulnerable to SQL injection (F5 demo surface).
      operationId: getToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              alice:
                summary: Alice (demo user)
                value: {username: alice, password: alice123}
              sqli:
                summary: SQL injection bypass
                value: {username: "' OR '1'='1' --", password: x}
      responses:
        "200":
          description: JWT issued successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Database error (verbose SQL returned -- intentional)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/login:
    post:
      tags: [Authentication]
      summary: Browser login (session cookie + JWT)
      description: |
        Authenticates the user, sets a Flask session cookie (for browser SPA
        navigation), and also returns a JWT in the response body.
        NOTE: Intentionally vulnerable to SQL injection (F5 demo surface).
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              thomas:
                summary: Thomas (demo user)
                value: {username: thomas, password: thomas123}
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Database error (verbose SQL returned -- intentional)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/me:
    get:
      tags: [Users]
      summary: Get the authenticated user profile
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/users:
    get:
      tags: [Users]
      summary: List all users
      description: Returns the full user directory.
      operationId: listUsers
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Array of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/users/{user_id}:
    get:
      tags: [Users]
      summary: Get a user profile by ID
      description: |
        Returns the full profile of the user identified by `{user_id}`.

        ⚠️  **INTENTIONALLY VULNERABLE TO BOLA (OWASP API1:2023 – Broken Object Level Authorization)**

        The endpoint only checks that the caller is *authenticated* (valid JWT).
        It does **not** verify that the requested `user_id` matches the authenticated
        user's own ID.  Any valid token holder can therefore enumerate all user
        profiles — including **plain-text passwords** — by iterating the integer ID:

        ```
        GET /api/users/1   → Alice's profile + password
        GET /api/users/2   → Thomas's profile + password
        GET /api/users/3   → Sophie's profile + password
        GET /api/users/4   → Lucas's profile + password
        ```

        This is the classic BOLA/IDOR pattern demonstrated for F5 WAF/API security
        demos. A secure implementation would return 403 when `user_id ≠ caller_id`.
      operationId: getUserById
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: Numeric ID of the target user (BOLA — no ownership check performed)
          schema:
            type: integer
            example: 2
      responses:
        "200":
          description: User profile (including plain-text password — intentional vuln)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/User"
                  - type: object
                    properties:
                      password:
                        type: string
                        description: Plain-text password exposed via BOLA
                        example: thomas123
              examples:
                alice:
                  summary: Alice (id=1)
                  value:
                    id: 1
                    name: Alice
                    surname: Moreau
                    email: alice.moreau@arcadiafinance.com
                    phone: "+33 6 12 34 56 78"
                    username: alice
                    password: alice123
                thomas:
                  summary: Thomas (id=2)
                  value:
                    id: 2
                    name: Thomas
                    surname: Lefebvre
                    email: thomas.lefebvre@arcadiafinance.com
                    phone: "+33 6 23 45 67 89"
                    username: thomas
                    password: thomas123
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/accounts:
    get:
      tags: [Accounts]
      summary: List bank accounts for the authenticated user
      operationId: listAccounts
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: query
          required: false
          description: Override the user whose accounts are returned. Defaults to the authenticated user.
          schema:
            type: integer
            example: 2
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
              example:
                - id: 1
                  account_number: FR7601234001001
                  type: checking
                  balance: 12450.75
                  currency: EUR
                  created_at: "2026-01-15T09:00:00"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/transfer:
    post:
      tags: [Transfers]
      summary: Execute a money transfer
      description: >
        Debits the source account and credits the destination account atomically.
        Proxied internally to the transfer-service (port 8081).
      operationId: createTransfer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
            example:
              from_account: FR7601234001001
              to_account: FR7601234002001
              amount: 250.00
              note: "Invoice #42"
      responses:
        "200":
          description: Transfer completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
        "400":
          description: Missing or invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Source or destination account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Database error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Transfer service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/transfers:
    get:
      tags: [Transfers]
      summary: Get transfer history for an account
      description: |
        Returns all transfers where the given account is sender or recipient,
        ordered by date descending (last 50 records).
        Proxied internally to the transfer-service (port 8081).
        NOTE: Intentionally vulnerable to SQL injection on the account param.
      operationId: listTransfers
      security:
        - BearerAuth: []
      parameters:
        - name: account
          in: query
          required: true
          description: Account number to retrieve history for
          schema:
            type: string
            example: FR7601234001001
      responses:
        "200":
          description: List of transfer records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransferRecord"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Database error (verbose SQL returned -- intentional)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Transfer service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...